Fix #1d - Rename the shared folder
*Renamed shared_temp folder to shared
