+++ src/app/app.module.ts
@@ -8,10 +8,10 @@ import { AppComponent } from './app.component';
 import { HeaderComponent } from './header/header.component';
 import { HomeComponent } from './home/home.component';
 
-import { LoginService } from '../app/shared_temp/login.service';
+import { LoginService } from '../app/shared/login.service';
 import { LoginComponent } from './login/login.component';
 
-import { UserService } from '../app/shared_temp/user.service';
+import { UserService } from '../app/shared/user.service';
 import { UsersComponent } from './users/users.component';
 import { UserDetailsComponent } from './users/user-details/user-details.component';
 import { UserEditComponent } from './users/user-details/user-edit/user-edit.component';
