+++ src/app/local-storage/local-storage.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { AlertService } from '../_alert';
-import { UserService } from '../shared_temp/user.service';
-import { TaskService } from '../shared_temp/task.service';
-import { ProjectService } from '../shared_temp/project.service';
+import { UserService } from '../shared/user.service';
+import { TaskService } from '../shared/task.service';
+import { ProjectService } from '../shared/project.service';
 
 
 @Component({
